+++ packages/javascript_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- Adds more exceptions that extend [JavaScriptPlatformException]
+- Updates existing [JavaScriptExecutionException] to extend [JavaScriptPlatformException]
+
 # 1.0.0+1
 
 - Update name of main package from `javascript` to `javascript_flutter`